all: make bun.FileDescriptor a unique type
# Conflicts:
#	src/bun.js/api/bun/socket.zig
#	src/bun.js/api/bun/spawn.zig
#	src/deps/uws.zig

+++ src/io/io.zig
@@ -770,7 +770,7 @@ pub const Poll = struct {
 
             kqueue_event.* = switch (comptime action) {
                 .readable => .{
-                    .ident = @as(u64, @intCast(fd)),
+                    .ident = @as(u64, @intCast(fd.int())),
                     .filter = std.os.system.EVFILT_READ,
                     .data = 0,
                     .fflags = 0,
@@ -779,7 +779,7 @@ pub const Poll = struct {
                     .ext = .{ generation_number, 0 },
                 },
                 .writable => .{
-                    .ident = @as(u64, @intCast(fd)),
+                    .ident = @as(u64, @intCast(fd.int())),
                     .filter = std.os.system.EVFILT_WRITE,
                     .data = 0,
                     .fflags = 0,
@@ -788,7 +788,7 @@ pub const Poll = struct {
                     .ext = .{ generation_number, 0 },
                 },
                 .cancel => if (poll.flags.contains(.poll_readable)) .{
-                    .ident = @as(u64, @intCast(fd)),
+                    .ident = @as(u64, @intCast(fd.int())),
                     .filter = std.os.system.EVFILT_READ,
                     .data = 0,
                     .fflags = 0,
@@ -796,7 +796,7 @@ pub const Poll = struct {
                     .flags = std.c.EV_DELETE,
                     .ext = .{ poll.generation_number, 0 },
                 } else if (poll.flags.contains(.poll_writable)) .{
-                    .ident = @as(u64, @intCast(fd)),
+                    .ident = @as(u64, @intCast(fd.int())),
                     .filter = std.os.system.EVFILT_WRITE,
                     .data = 0,
                     .fflags = 0,
@@ -814,7 +814,7 @@ pub const Poll = struct {
         _ = linux.epoll_ctl(
             @intCast(watcher_fd),
             linux.EPOLL.CTL_DEL,
-            @intCast(fd),
+            @intCast(fd.int()),
             null,
         );
         this.flags.remove(.was_ever_registered);
@@ -900,7 +900,7 @@ pub const Poll = struct {
             const ctl = linux.epoll_ctl(
                 watcher_fd,
                 op,
-                @intCast(fd),
+                @intCast(fd.int()),
                 &event,
             );
             this.flags.insert(.registered);
